pyproject.toml
[project]
name = "the-bureaucracy-buster-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "edvardvet",email = "cstteam3@gmail.com"}
]
readme = "README.md"
dependencies = [
    "pymupdf (>=1.26.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-core (>=0.3.65,<0.4.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "langchain-community (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.23,<0.4.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "openai (>=1.86.0,<2.0.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "python-telegram-bot (>=22.1,<23.0)"
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
